feat: 운동 무게 소수점 가능하도록 변경

+++ src/main/kotlin/com/jagaldol/behind/fitness/workout/session/dto/SessionResponse.kt
@@ -52,7 +52,7 @@ class SessionResponse {
 
                 data class GetSetDto(
                     val id: Long,
-                    val weight: Int,
+                    val weight: Double,
                     val count: Int,
                     val countUnit: String
                 ) {
@@ -107,7 +107,7 @@ class SessionResponse {
 
                 data class GetSetDto(
                     val id: Long,
-                    val weight: Int,
+                    val weight: Double,
                     val count: Int,
                     val countUnit: String
                 ) {
